Multiplayer Changes
Player can now see enemy units as well as their constant movements.
Needs a bit of work to be efficient and properly work but base idea is
there. MultiplayerGamePlayScreen is no longer a thing, all games be it
single or mp take place within the GameplayScreen class and it simply
updates the players playing regardless of whether they are netplayers or
ai. Basically, it is streamlined and we can now have ai players in the
same room as mp or "bots".