Econo Patch Number 5
Economies can only add up to a max of 100% in a single factory, exporter
or market. On top of that, a factory cannot output less in a specific
economy than what the exporters are currently set to. And exporters can
only export as much as the factories are creating. These numbers are
visually displayed. Big update.