Group Unit Movement Fix V1
Good news, many types of formations are now able to move. However, there
are still problems with units on the same column as the leader. Am
working on this, but so far tests are considerably more positive than
any previous ones. The moved units even appear to have a buildable tile
underneath them.