Issue #36 Squad Leader Problems Rest of Fix
Issues reported with squad leader functionality have been fixed or i've
been unable to recreate them, as is the case with units counting down
their numbers and dying.
Fixed: Only one unit can be the leader of a group. If one leader tries
to join the squad of another leader/new leader, the current one and all
of its following units are absorbed.