Issue #36 Squad Leader Problems Closed
I'm going to call this one closed. Just fixed the problem of following
units colliding with their own units outside of their group. The code
told them to do battle with their own people if they encountered anyone
so it makes sense that they were dying as if they were in combat.